+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 **Implemented enhancements:**
 - added multiple options [`x-powered-by`, `routing.strict`, `routing.sensitive`, `json.escape`]
-- added multiple settings [`enc`, `url`, `port`, `json.replacer`, `json.spaces`, `query.parser`]
+- added multiple settings [`env`, `url`, `port`, `json.replacer`, `json.spaces`, `query.parser`]
 - added multiple model relations [`hasOne`, `hasMany`]
 - improved usage